+++ tileview/src/main/java/com/qozix/tileview/detail/DetailLevelEventListener.java
@@ -2,5 +2,5 @@ package com.qozix.tileview.detail;
 
 public interface DetailLevelEventListener {
 	public void onDetailLevelChanged();
-	public void onDetailScaleChanged( double scale );
+	public void onDetailScaleChanged( float scale );
 }
